finish styling
