+++ src/styles/components/_contact-form.scss
@@ -1,4 +1,5 @@
 .contacts__form {
+  width: 80%;
   display: flex;
   flex-direction: column;
 
@@ -9,4 +10,7 @@
   align-self: flex-start;
   font-size: 1.7rem;
   justify-self: flex-end;
+  &:hover {
+    filter: brightness(1.1);
+  }
 }
