+++ src/styles/components/_form.scss
@@ -36,4 +36,8 @@
   border-radius: $border-radius;
   border: none;
   background-color: $color-primary;
+  transition: all 0.2s;
+  &:hover {
+    filter: brightness(1.1);
+  }
 }
