+++ src/styles/pages/_contacts.scss
@@ -4,12 +4,14 @@
   width: 80%;
   max-width: 800px;
 
-  display: flex;
-  justify-content: space-between;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(25rem, 1fr));
   gap: 3rem;
-  flex-wrap: wrap;
 }
 
 .contacts__exit-btn {
-  flex-basis: 5%;
+  grid-column: 2/3;
+  &:hover {
+    filter: brightness(1.1);
+  }
 }
